change isvalid to actually do stuff

+++ Market-Trading-System/src/main/java/ServiceLayer/Store/StoreManagementService.java
@@ -12,6 +12,7 @@ import javax.ws.rs.core.Response;
 import java.util.List;
 
 public class StoreManagementService implements IStoreManagementService{
+    String USER_NOT_VALID = "Authentication failed";
     private static StoreManagementService instance;
     private final IStoreFacade storeFacade;
     private JwtService jwtService;
@@ -33,9 +34,11 @@ public class StoreManagementService implements IStoreManagementService{
     public Response createStore(String founderToken, String storeName, String storeDescription, IRepository<Long, Discount> repository) {
         try {
             String userName = jwtService.extractUsername(founderToken);
-            jwtService.isValid(founderToken, userDetailsService.loadUserByUsername(userName));
-            storeFacade.createStore(userName,storeName,storeDescription,repository);
-            return Response.ok().build();
+            if(jwtService.isValid(founderToken, userDetailsService.loadUserByUsername(userName))) {
+                storeFacade.createStore(userName, storeName, storeDescription, repository);
+                return Response.ok().build();
+            }
+            else return Response.status(500).entity(USER_NOT_VALID).build();
         }
         catch (Exception ex){
             return Response.status(500).entity(ex.getMessage()).build();
@@ -46,9 +49,11 @@ public class StoreManagementService implements IStoreManagementService{
     public Response addItemToStore(String token, long storeId, String itemName, String description ,double itemPrice, int stockAmount, List<String> categories) {
         try {
             String userName = jwtService.extractUsername(token);
-            jwtService.isValid(token, userDetailsService.loadUserByUsername(userName));
-            storeFacade.addItemToStore(userName,storeId,itemName,itemPrice,stockAmount, description,categories);
-            return Response.ok().build();
+            if(jwtService.isValid(token, userDetailsService.loadUserByUsername(userName))) {
+                storeFacade.addItemToStore(userName, storeId, itemName, itemPrice, stockAmount, description, categories);
+                return Response.ok().build();
+            }
+            else return Response.status(500).entity(USER_NOT_VALID).build();
         }
         catch (Exception ex){
             return Response.status(500).entity(ex.getMessage()).build();
@@ -59,9 +64,11 @@ public class StoreManagementService implements IStoreManagementService{
     public Response updateItem(String token, long storeId, long itemId, String newName, double newPrice, int newAmount) {
         try {
             String userName = jwtService.extractUsername(token);
-            jwtService.isValid(token, userDetailsService.loadUserByUsername(userName));
-            storeFacade.updateItem(userName,storeId,itemId,newName,newPrice, newAmount);
-            return Response.ok().build();
+            if(jwtService.isValid(token, userDetailsService.loadUserByUsername(userName))){
+                storeFacade.updateItem(userName,storeId,itemId,newName,newPrice, newAmount);
+                return Response.ok().build();
+            }
+            else return Response.status(500).entity(USER_NOT_VALID).build();
         }
         catch (Exception ex){
             return Response.status(500).entity(ex.getMessage()).build();
@@ -73,9 +80,11 @@ public class StoreManagementService implements IStoreManagementService{
     public Response deleteItem(String token, long storeId, long itemId) {
         try {
             String userName = jwtService.extractUsername(token);
-            jwtService.isValid(token, userDetailsService.loadUserByUsername(userName));
-            storeFacade.deleteItem(userName, storeId, itemId);
-            return Response.ok().build();
+            if(jwtService.isValid(token, userDetailsService.loadUserByUsername(userName))) {
+                storeFacade.deleteItem(userName, storeId, itemId);
+                return Response.ok().build();
+            }
+            else return Response.status(500).entity(USER_NOT_VALID).build();
         }
         catch (Exception ex){
             return Response.status(500).entity(ex.getMessage()).build();
@@ -86,9 +95,11 @@ public class StoreManagementService implements IStoreManagementService{
     public Response changeStorePolicy(String token, long storeId) {
         try {
             String userName = jwtService.extractUsername(token);
-            jwtService.isValid(token, userDetailsService.loadUserByUsername(userName));
-            storeFacade.changeStorePolicy(userName, storeId);
-            return Response.ok().build();
+            if(jwtService.isValid(token, userDetailsService.loadUserByUsername(userName))){
+                storeFacade.changeStorePolicy(userName, storeId);
+                return Response.ok().build();
+            }
+            else return Response.status(500).entity(USER_NOT_VALID).build();
         }
         catch (Exception ex){
             return Response.status(500).entity(ex.getMessage()).build();
@@ -99,9 +110,11 @@ public class StoreManagementService implements IStoreManagementService{
     public Response changeDiscountType(String token, long storeId, String newType) {
         try {
             String userName = jwtService.extractUsername(token);
-            jwtService.isValid(token, userDetailsService.loadUserByUsername(userName));
-            storeFacade.changeDiscountType(userName ,storeId, newType);
-            return Response.ok().build();
+            if(jwtService.isValid(token, userDetailsService.loadUserByUsername(userName))) {
+                storeFacade.changeDiscountType(userName ,storeId, newType);
+                return Response.ok().build();
+            }
+            else return Response.status(500).entity(USER_NOT_VALID).build();
         }
         catch (Exception ex){
             return Response.status(500).entity(ex.getMessage()).build();
@@ -112,9 +125,11 @@ public class StoreManagementService implements IStoreManagementService{
     public Response removeStore(String token, long storeId) {
         try {
             String userName = jwtService.extractUsername(token);
-            jwtService.isValid(token, userDetailsService.loadUserByUsername(userName));
-            storeFacade.removeStore(userName ,storeId);
-            return Response.ok().build();
+            if(jwtService.isValid(token, userDetailsService.loadUserByUsername(userName))) {
+                storeFacade.removeStore(userName, storeId);
+                return Response.ok().build();
+            }
+            else return Response.status(500).entity(USER_NOT_VALID).build();
         }
         catch (Exception ex){
             return Response.status(500).entity(ex.getMessage()).build();
@@ -125,8 +140,10 @@ public class StoreManagementService implements IStoreManagementService{
     public Response viewManagmentInfo(String token, Long storeId) {
         try {
             String userName = jwtService.extractUsername(token);
-            jwtService.isValid(token, userDetailsService.loadUserByUsername(userName));
-            return Response.ok().entity(storeFacade.viewStoreManagementInfo(token, storeId)).build();
+            if(jwtService.isValid(token, userDetailsService.loadUserByUsername(userName))) {
+                return Response.ok().entity(storeFacade.viewStoreManagementInfo(token, storeId)).build();
+            }
+            else return Response.status(500).entity(USER_NOT_VALID).build();
         }
         catch (Exception ex){
             return Response.status(500).entity(ex.getMessage()).build();
@@ -137,8 +154,10 @@ public class StoreManagementService implements IStoreManagementService{
     public Response viewInventory(String token, Long storeId) {
         try {
             String userName = jwtService.extractUsername(token);
-            jwtService.isValid(token, userDetailsService.loadUserByUsername(userName));
-            return Response.ok().entity(storeFacade.viewInventoryByStoreOwner(userName, storeId)).build();
+            if (jwtService.isValid(token, userDetailsService.loadUserByUsername(userName))) {
+                return Response.ok().entity(storeFacade.viewInventoryByStoreOwner(userName, storeId)).build();
+            }
+            else return Response.status(500).entity(USER_NOT_VALID).build();
         }
         catch (Exception ex){
             return Response.status(500).entity(ex.getMessage()).build();
@@ -149,8 +168,10 @@ public class StoreManagementService implements IStoreManagementService{
     public Response viewPurchasesHistory(String token, Long storeId) {
         try {
             String userName = jwtService.extractUsername(token);
-            jwtService.isValid(token, userDetailsService.loadUserByUsername(userName));
+            if(jwtService.isValid(token, userDetailsService.loadUserByUsername(userName))){
             return Response.ok().entity(storeFacade.viewPurchaseHistory(userName, storeId)).build();
+            }
+            else return Response.status(500).entity(USER_NOT_VALID).build();
         }
         catch (Exception e){
             throw new RuntimeException(e.getMessage());
@@ -161,9 +182,11 @@ public class StoreManagementService implements IStoreManagementService{
     public Response assignStoreOwner(String token, long storeId, String newOwnerId) {
         try {
             String userName = jwtService.extractUsername(token);
-            jwtService.isValid(token, userDetailsService.loadUserByUsername(userName));
-            storeFacade.assignStoreOwner(userName, storeId, newOwnerId);
-            return Response.ok().build();
+            if(jwtService.isValid(token, userDetailsService.loadUserByUsername(userName))) {
+                storeFacade.assignStoreOwner(userName, storeId, newOwnerId);
+                return Response.ok().build();
+            }
+            else return Response.status(500).entity(USER_NOT_VALID).build();
         }
         catch (Exception ex){
             return Response.status(500).entity(ex.getMessage()).build();
@@ -174,9 +197,11 @@ public class StoreManagementService implements IStoreManagementService{
     public Response assignStoreManager(String token, long storeId, String newManagerI, List<String> permissions) {
         try {
             String userName = jwtService.extractUsername(token);
-            jwtService.isValid(token, userDetailsService.loadUserByUsername(userName));
-            storeFacade.assignStoreManager(userName, storeId, newManagerI, permissions);
-            return Response.ok().build();
+            if(jwtService.isValid(token, userDetailsService.loadUserByUsername(userName))) {
+                storeFacade.assignStoreManager(userName, storeId, newManagerI, permissions);
+                return Response.ok().build();
+            }
+            else return Response.status(500).entity(USER_NOT_VALID).build();
         }
         catch (Exception ex){
             return Response.status(500).entity(ex.getMessage()).build();
